Change handling of IntegerNode and BinaryNode constructors
Removed the use of `std::variant`. Perhaps with templating in the
future we can reduce the # of constructors.

+++ tests/cpp/nodes/test_numbers.cpp
@@ -252,7 +252,7 @@ TEST_CASE("BinaryNode") {
     }
     GIVEN("Binary node with index-wise bounds") {
         auto bnode_ptr = graph.emplace_node<dwave::optimization::BinaryNode>(
-                3, std::vector<ssize_t>{-1, 0, 1}, std::vector<ssize_t>{2, 1, 1});
+                3, std::vector<double>{-1, 0, 1}, std::vector<double>{2, 1, 1});
 
         THEN("The shape, max, min, and index-wise bounds are correct") {
             CHECK(bnode_ptr->size() == 3);
@@ -328,12 +328,12 @@ TEST_CASE("BinaryNode") {
 
     GIVEN("Binary node with invalid index-wise lower bounds at index 0") {
         REQUIRE_THROWS(graph.emplace_node<dwave::optimization::BinaryNode>(
-                2, std::vector<ssize_t>{2, 0}, std::vector<ssize_t>{1, 1}));
+                2, std::vector<double>{2, 0}, std::vector<double>{1, 1}));
     }
 
     GIVEN("Binary node with invalid index-wise upper bounds at index 1") {
         REQUIRE_THROWS(graph.emplace_node<dwave::optimization::BinaryNode>(
-                2, std::vector<ssize_t>{0, 0}, std::vector<ssize_t>{1, -1}));
+                2, std::vector<double>{0, 0}, std::vector<double>{1, -1}));
     }
 }
 
@@ -401,7 +401,7 @@ TEST_CASE("IntegerNode") {
 
     GIVEN("Integer node with index-wise bounds") {
         auto inode_ptr = graph.emplace_node<dwave::optimization::IntegerNode>(
-                3, std::vector<ssize_t>{-1, 3, 5}, std::vector<ssize_t>{1, 6, 7});
+                3, std::vector<double>{-1, 3, 5}, std::vector<double>{1, 6, 7});
 
         THEN("The shape, max, min, and index-wise bounds are correct") {
             CHECK(inode_ptr->size() == 3);
@@ -447,7 +447,7 @@ TEST_CASE("IntegerNode") {
 
     GIVEN("Integer node with index-wise upper bound and general integer lower bound") {
         auto inode_ptr = graph.emplace_node<dwave::optimization::IntegerNode>(
-                2, 10, std::vector<ssize_t>{20, 10});
+                2, 10, std::vector<double>{20, 10});
 
         THEN("The max, min, and index-wise bounds are correct") {
             CHECK(inode_ptr->max() == 20.0);
@@ -462,7 +462,7 @@ TEST_CASE("IntegerNode") {
 
     GIVEN("Integer node with invalid index-wise bounds at index 0") {
         REQUIRE_THROWS(graph.emplace_node<dwave::optimization::IntegerNode>(
-                2, std::vector<ssize_t>{19, 12}, std::vector<ssize_t>{20, 11}));
+                2, std::vector<double>{19, 12}, std::vector<double>{20, 11}));
     }
 
     GIVEN("An Integer Node representing an 1d array of 10 elements with lower bound -10") {
